fix types to compile on windows

+++ include/packet.hh
@@ -51,7 +51,7 @@ public:
   // std::vector<uint8_t> pop( PacketTag tag );
 
   uint8_t &data() { return buffer.at(headerSize); }
-  int size() { return buffer.size() - headerSize; }
+  size_t size() { return buffer.size() - headerSize; }
   void resize(int size) { buffer.resize(headerSize + size); }
 
   std::vector<uint8_t> buffer;
